Move component to components-ui
